+++ server/config/routes.rb
@@ -10,6 +10,7 @@ Rails.application.routes.draw do
       resources :tickets
       resources :events
       resources :venues
+      resources :users
     end
   end
   # Define your application routes per the DSL in https://guides.rubyonrails.org/routing.html
